+++ resources/views/fields/select2_from_ajax_multiple.blade.php
@@ -2,7 +2,7 @@
 @php
     $connected_entity = new $field['model'];
     $connected_entity_key_name = $connected_entity->getKeyName();
-    $old_value = old($field['name']) ? old($field['name']) : (isset($field['value']) ? $field['value'] : (isset($field['default']) ? $field['default'] : false ));
+    $old_value = old(square_brackets_to_dots($field['name'])) ?? $field['value'] ?? $field['default'] ?? false;
 @endphp
 
 <div @include('crud::inc.field_wrapper_attributes') >
@@ -43,7 +43,7 @@
 {{-- ########################################## --}}
 {{-- Extra CSS and JS for this particular field --}}
 {{-- If a field type is shown multiple times on a form, the CSS and JS will only be loaded once --}}
-@if ($crud->checkIfFieldIsFirstOfItsType($field, $fields))
+@if ($crud->checkIfFieldIsFirstOfItsType($field))
 
     {{-- FIELD CSS - will be loaded in the after_styles section --}}
     @push('crud_fields_styles')
@@ -69,6 +69,8 @@
         jQuery(document).ready(function ($) {
             // trigger select2 for each untriggered select2 box
             $("#select2_ajax_multiple_{{ $field['name'] }}").each(function (i, obj) {
+                var form = $(obj).closest('form');
+
                 if (!$(obj).hasClass("select2-hidden-accessible")) {
                     $(obj).select2({
                         theme: 'bootstrap',
@@ -77,13 +79,15 @@
                         minimumInputLength: "{{ $field['minimum_input_length'] }}",
                         ajax: {
                             url: "/{{ ltrim($field['data_source'] ?? $crud->getRoute().'/ajax', '/') }}",
+                            type: '{{ $field['method'] ?? 'GET' }}',
                             dataType: 'json',
                             quietMillis: 250,
                             data: function (params) {
                                 return {
                                     q: params.term, // search term
                                     field: "{{ $field['name'] }}",
-                                    page: params.page
+                                    page: params.page,
+                                    form: form.serializeArray() // all other form inputs
                                 };
                             },
                             processResults: function (data, params) {
@@ -104,6 +108,14 @@
                     });
                 }
             });
+
+            @if (isset($field['dependencies']))
+                @foreach (array_wrap($field['dependencies']) as $dependency)
+                    $('input[name={{ $dependency }}], select[name={{ $dependency }}], checkbox[name={{ $dependency }}], radio[name={{ $dependency }}], textarea[name={{ $dependency }}]').change(function () {
+                        $("#select2_ajax_multiple_{{ $field['name'] }}").val(null).trigger("change");
+                    });
+                @endforeach
+            @endif
         });
     </script>
 @endpush